cleanup + try make cli log without .env
